social_media_api/requirements.txt
argcomplete==3.5.0
asgiref==3.8.1
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1
django-extensions==3.2.3
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
durationpy==0.7
hjson==3.1.0
idna==3.10
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
placebo==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
zappa==0.59.0